services/notification-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
boto3==1.29.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Database
pymongo==4.5.0
motor==3.3.1

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1